+++ changelog.md
@@ -14,7 +14,8 @@
       - `Commands` Echo #commands to the screen
     - Command line:
       - `Auto size to contents` Resize command input height based on contents
-      - `Word wrap` Wrap long lines in command input    
+      - `Word wrap` Wrap long lines in command input
+      - `Scrollbars` Add scrollbars when needed to command input
   - Code editor:
     - Add right click spell check support for systems that support it
     - Add basic spell checking